pyproject.toml
[project]
name = "TranscriptionTool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "imageio-ffmpeg>=0.6.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

